nes-datagen: rename NCLP to cursor-jump throughout
Drop the NCLP abbreviation in favor of the more descriptive
'cursor-jump' name already used in the production xtab provider.
 cursorJumpPromptStep,
 cursorJumpResponseStep), the capture request ids,
and all surrounding doc comments / test descriptions.
