+++ extensions/copilot/test/pipeline/cursorJump/cursorJumpResponseStep.ts
